requirements.txt
ultralytics>=8.3.0
roboflow>=1.1.34
python-dotenv>=1.0.1

opencv-python>=4.10.0
numpy>=1.26.0
matplotlib>=3.8.0

torch>=2.1.0
torchvision>=0.16.0

onnx>=1.12.0,<1.18.0
onnxruntime>=1.15.0
onnxslim>=0.1.59
coremltools>=7.0  ; platform_system == "Darwin"
tensorflow>=2.15.0 ; platform_system != "Darwin"
